final_app/requirements.txt
django==3.1.7
djangorestframework==3.14
django-mysql==3.11.1
django-cors-headers==3.7.0
mysqlclient==2.0.3
PyJwt==2.0.1
Faker==6.6.1
django-redis==4.12.1
stripe==2.56.0
python-dotenv



# sudo apt-get install default-libmysqlclient-dev for installing mysqlclient